+++ AS/deletestaffview.php
@@ -26,6 +26,10 @@
 </head>
 
 <body>
+<?php
+	session_start();
+	$role = $_SESSION['role'];
+	?>
     <!-- ============================================================== -->
     <!-- Preloader - style you can find in spinners.css -->
     <!-- ============================================================== -->
@@ -118,9 +122,13 @@
                         <li class="sidebar-item"> <a class="sidebar-link waves-effect waves-dark sidebar-link"
                                 href="feedbackView.php" aria-expanded="false"><i class="mdi mdi-file"></i><span
                                     class="hide-menu">Feedback</span></a></li>
-                        <li class="sidebar-item"> <a class="sidebar-link waves-effect waves-dark sidebar-link"
-                                href="deletestaffview.php" aria-expanded="false"><i class="mdi mdi-face"></i><span
-                                    class="hide-menu">Staff</span></a></li>
+						<?php
+						if($role === "admin"){
+                        echo "<li class='sidebar-item'> <a class='sidebar-link waves-effect waves-dark sidebar-link'
+                                href='deletestaffview.php' aria-expanded='false'><i class='mdi mdi-face'></i><span
+                                    class='hide-menu'>Staff</span></a></li>";
+						}
+						?>
 						<li class="sidebar-item"> <a class="sidebar-link waves-effect waves-dark sidebar-link"
                                 href="deletecustomerview.php" aria-expanded="false"><i class="mdi mdi-face"></i><span
                                     class="hide-menu">Customer</span></a></li>
@@ -150,12 +158,12 @@
             <div class="page-breadcrumb">
                 <div class="row align-items-center">
                     <div class="col-5">
-                        <h4 class="page-title">Feedback</h4>
+                        <h4 class="page-title">Staff</h4>
                         <div class="d-flex align-items-center">
                             <nav aria-label="breadcrumb">
                                 <ol class="breadcrumb">
                                     <li class="breadcrumb-item"><a href="index.html">Home</a></li>
-                                    <li class="breadcrumb-item active" aria-current="page">Feedback</li>
+                                    <li class="breadcrumb-item active" aria-current="page">Staff</li>
                                 </ol>
                             </nav>
                         </div>
